requirements.txt
asgiref==3.2.7
astroid @ file:///tmp/build/80754af9/astroid_1592495912941/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1592412889138/work
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.3
django-rest-knox==4.1.0
djangorestframework==3.11.0
httpie==2.2.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint @ file:///tmp/build/80754af9/pylint_1592496053748/work
pylint-django==2.0.15
pylint-plugin-utils==0.6
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
urllib3==1.25.9
wrapt==1.11.2
